Sometimes also print a usage message on time parse errors
One way to have waituntil's wait time fail to parse is to forget the
usage of waituntil and try something like 'waituntil 10:23 cmd arg'.
Now we also print a usage message if the time fails to parse and there
is more than one (non-flag) argument, as a hint about what may be
wrong.